Do a full clone so that we have enough information to determine the release (#29)
This PR ensures that in the release workflow we do a full clone, so that setuptools_scm has enough information to compute a version number.

We don't strictly need this if we only ever expect to be making releases from a tagged commit (which seems likely), but the failure mode if this is left out is that we compute a wrong version. So it seems better to include it. If some future version of setuptools_scm makes this an error, it should be safe to remove this.

See pypa/setuptools_scm#480 for more discussion.