Don't mention license twice (#30)
The use of license = {file = 'LICENSE'} results in significant ugliness in the sidebar on PyPI. We don't appear to need it, since the license is already specified in the trove classifiers. This PR removes that entry from the pyproject.toml (and also does a drive-by cleanup of the authors field).

(cherry picked from commit df42c4386c31553a5c7aa4fa5b7a99add62bff31)